requirements.txt
streamlit==1.24.1
pandas==2.0.1
streamlit-drawable-canvas==0.9.2
Pillow==9.5.0
PyPDF2==3.0.1
reportlab==3.6.13
python-dateutil==2.8.2
xlsxwriter==3.0.2
pytz==2021.1
numpy==1.24.3